requirements.txt
PyYAML>=6.0
pytest>=8.0
requests>=2.31
